Add count! fn
This allows retrieving the count from count queries. It's a bit of a
hack since it's the exact same underlying rest method as query!, just a
different return type and without the support for paginated results.
(Arguably that last bit is a feature since paginated results make no
sense for count queries.)