Add take-action! implementation to memory client
The memory client can be reified with a map of action names
to functions that can be passed in when building the client.
Then take-action! can be called with an action name along
with a function that will operate on the client in memory
along with arguments passed to that function.